Fixing and consolidating types
