+++ server/components/AppContext.tsx
@@ -1,7 +1,13 @@
-import { AppState, AppStateCtx, Attempt, PromptRequest } from "@/interfaces/ui";
 import { createContext, useState, FC, ReactNode, useEffect } from "react";
 import { useSession } from "@supabase/auth-helpers-react";
-import { PromptResponse } from "@/lib/playground";
+import {
+  DetectResponse,
+  PromptResponse,
+  AppState,
+  AppStateCtx,
+  Attempt,
+  PromptRequest,
+} from "@/types/types";
 
 const initState = {
   apikey: "",
