+++ server/lib/account-helpers.ts
@@ -1,8 +1,8 @@
-import { AppState } from "@/interfaces/ui";
+import { AppState } from "@/types/types";
 import generateApiKey from "@/utils/apikeys";
 import { supabaseAdminClient } from "@/lib/supabase";
 import { User } from "@supabase/supabase-js";
-import { PromptResponse } from "./playground";
+import { PromptResponse } from "@/types/types";
 
 export const createNewAccountInDb = async (user: any): Promise<AppState> => {
   const { data, error } = await supabaseAdminClient
