+++ server/pages/api/account/index.ts
@@ -7,7 +7,7 @@ import {
   getUserStats,
   refreshUserApikeyInDb,
 } from "@/lib/account-helpers";
-import { AppState } from "@/interfaces/ui";
+import { AppState } from "@/types/types";
 type MiddlewareCallback = (result: any) => void;
 
 const cors = Cors({
